+++ .eslintrc
@@ -1,0 +1,49 @@
+{
+  "parser": "babel-eslint",
+  "extends": [
+    "airbnb",
+    "plugin:jest/recommended"
+  ],
+  "plugins": [
+    "jest",
+    "mocha",
+    "react"
+  ],
+  "globals": {
+    "configureMockStore": true,
+    "expect": true,
+    "snapshot": true
+  },
+  "env": {
+    "browser": true,
+    "node": true,
+    "jest/globals": true
+  },
+  "rules": {
+    "arrow-body-style": [0],
+    "comma-dangle": [2, "never"],
+    "consistent-return": [0],
+    "global-require": [0],
+    "import/no-named-as-default": 0,
+    "import/prefer-default-export": [0],
+    "import/no-extraneous-dependencies": [0],
+    "import/no-dynamic-require": [0],
+    "import/order": [0],
+    "jest/valid-expect": "error",
+    "max-len": [2, 120, 4, {"ignoreUrls": true}],
+    "mocha/no-exclusive-tests": [2],
+    "newline-per-chained-call": [0],
+    "no-buffer-constructor": [0],
+    "no-case-declarations": [0],
+    "no-invalid-this": 0,
+    "no-param-reassign": [2, { "props": false }],
+    "no-plusplus": 0,
+    "no-underscore-dangle": [0],
+    "no-unused-vars": [2, { "args": "none" }],
+    "no-unused-expressions": [0],
+    "react/prefer-stateless-function": [0],
+    "react/jsx-first-prop-new-line": [0],
+    "react/jsx-filename-extension": [0],
+    "quote-props": [2, "as-needed", { "keywords": true, "unnecessary": false }],
+  }
+}
