+++ app/containers/App.test.js
@@ -1,12 +1,12 @@
 import React from 'react';
-import { App } from './App';
 import renderer from 'react-test-renderer';
+import { App } from './App';
 import { t } from '../../test/i18n/mocks';
 
 describe('App', () => {
-    it('renders', () => {
-        const component = renderer.create(<App t={t} />);
-        let tree = component.toJSON();
-        expect(tree).toMatchSnapshot();
-    });
+  it('renders', () => {
+    const component = renderer.create(<App t={t} />);
+    const tree = component.toJSON();
+    expect(tree).toMatchSnapshot();
+  });
 });
