+++ app/index.js
@@ -1,17 +1,17 @@
 import 'helix-ui/dist/css/helix-ui.css';
-import React from "react";
+import React from 'react';
 import { render } from 'react-dom';
-import '@webcomponents/webcomponentsjs/webcomponents-bundle.js';
-import App from "./containers/App.js";
+import App from './containers/App';
+import '@webcomponents/webcomponentsjs/webcomponents-bundle';
 import './i18n';
 
 const initializeJupiter = () => {
-    const content = (
-        <div id="jupiter-wrapper" className="u-flex-grow">
-            <App/>
-        </div>
-    );
-    render(content, document.getElementById("jupiter"));
+  const content = (
+    <div id="jupiter-wrapper" className="u-flex-grow">
+      <App />
+    </div>
+  );
+  render(content, document.getElementById('jupiter'));
 };
 
 initializeJupiter();
