+++ i18n/index.js
@@ -3,10 +3,10 @@ import commonEn from './en-US/common.json';
 import commonJa from './ja-JP/common.json';
 
 export const resources = {
-    'en-US': {
-        common: commonEn
-    },
-    'ja-JP': {
-        common: commonJa
-    }
+  'en-US': {
+    common: commonEn
+  },
+  'ja-JP': {
+    common: commonJa
+  }
 };
