+++ package.json
@@ -11,6 +11,7 @@
     "build": "webpack --config ./webpack.prod.js",
     "test": "jest",
     "test:coverage": "jest --coverage",
+    "lint": "eslint app lib test",
     "translations-scan": "i18next-scanner --config i18next-scanner.config.js"
   },
   "dependencies": {
@@ -32,6 +33,7 @@
     "@babel/plugin-proposal-class-properties": "^7.7.0",
     "@babel/preset-env": "^7.7.1",
     "@babel/preset-react": "^7.7.0",
+    "babel-eslint": "^10.0.3",
     "babel-loader": "^8.0.6",
     "clean-webpack-plugin": "^3.0.0",
     "copy-webpack-plugin": "^5.0.5",
@@ -39,6 +41,14 @@
     "enzyme": "^3.10.0",
     "enzyme-adapter-react-16": "^1.15.1",
     "eol": "^0.9.1",
+    "eslint": "^6.6.0",
+    "eslint-config-airbnb": "^18.0.1",
+    "eslint-plugin-babel": "^5.3.0",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-jest": "^23.0.3",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-mocha": "^6.2.1",
+    "eslint-plugin-react": "^7.16.0",
     "html-webpack-plugin": "^3.2.0",
     "i18next-scanner": "^2.10.3",
     "jest": "^24.9.0",
