+++ test/i18n/mocks.js
@@ -1,7 +1,7 @@
 import i18n from './setup';
 
 function i18nTMock() {
-    return i18n.t.bind(i18n);
+  return i18n.t.bind(i18n);
 }
 
 export const t = i18nTMock();
